Implement dotfile mapping includes
When a dotfile mapping includes an #include directive, it will load the
requested mapping file as well, and this can be done recursively to any
extent requested.  This is done to allow, for instance, nested dotfile
repositories to each have their own dotfile mappings, and the containing
repo's mapping can then chain to the nested mappings.